requirements.txt

flask
pandas
python-docx
openpyxl
